+++ src/Program.cs
@@ -103,6 +103,8 @@ public static class Program
             hostContext.Configuration.GetSection(nameof(ConfigOptions.Redis)));
         services.Configure<CircuitBreakerSettings>(
             hostContext.Configuration.GetSection(nameof(CircuitBreakerSettings)));
+        services.Configure<DbCredentialsFileSettings>(
+            hostContext.Configuration.GetSection(nameof(DbCredentialsFileSettings)));
     }
 
     private static void RegisterSerilog(IServiceCollection services)